Separation of jvm-specific code from SQLStringContextUpdateTest
